added login capability and made it so that each login has its own team roster they can fill out, also swapped the index page so that you are immediately prompted to register or login

+++ assignments/CourseProject/delete.php
@@ -1,14 +1,19 @@
 <?php
 // deletes a player from the database using the player id
-
-// database connection
+//database connection and user authentication
+require "includes/auth.php";
 require "includes/connect.php";
 
 // make sure an id was passed in the url
 $playerId = $_GET['id'];
 
 // sql query to delete player
-$sql = "DELETE FROM players WHERE player_id = :player_id";
+$userId = $_SESSION['user_id'];
+$sql = "DELETE FROM players WHERE player_id = :player_id AND user_id = :user_id";
+$stmt = $pdo->prepare($sql);
+$stmt->bindParam(':player_id', $playerId);
+$stmt->bindValue(':user_id', $userId, PDO::PARAM_INT);
+$stmt->execute();
 
 // prepare sql statement
 $stmt = $pdo->prepare($sql);
